Finish

+++ routes/items.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 let Item = require('../models/item.model');
-
+const List = require('../models/listTitles.model')
 
 router.route('/').get((req, res) => {
     
@@ -12,25 +12,45 @@ router.route('/').get((req, res) => {
 
 router.route('/').post((req, res) => {
     const itemValue = req.body.newItem;
-
+    const listName = req.body.list;
+    
     const newItem = new Item({name: itemValue});
-    newItem.save()
-    .then(() => console.log("Successfully added"));
 
-    res.redirect('/');
+    if (listName === 'Today'){
+        newItem.save()
+        .then(() => console.log("Successfully added"));
+        res.redirect('/');
+    } else {
+        List.findOne({title: listName}, function(err, foundList){
+            foundList.items.push(newItem);
+            foundList.save();
+            res.redirect("/"+ listName);
+        })
+    } 
+   
 });
 
 router.route('/delete').post((req, res) => {
     const checkedIt = req.body.checkbox;
-    Item.findByIdAndRemove({_id: checkedIt}, (err) => {
-        if(!err){
-            console.log("Successfully delete!");
-        }
-        res.redirect('/');
-    })
+    const listName = req.body.listName;
+
+    if (listName === "Today"){
+        Item.findByIdAndRemove({_id: checkedIt}, (err) => {
+            if(!err){
+                console.log("Successfully delete!");
+            }
+            res.redirect('/');
+        });
+    } else {
+        List.findOneAndUpdate({title: listName}, {$pull: {items: {_id: checkedIt}}}, function(err,  foundList){
+            if(!err){
+                res.redirect("/"+ listName);
+            }
+        })
+    }
+
     
-})
-
-
+    
+});
 
 module.exports = router;
